0.51.0: Bump mongochangestream to pick up refresh token improvements
Also, use `assertEventually` to make tests more reliable when latency is bad:

I'm currently in Brazil and experiencing poor latency, which is making the tests
fail the way that they're written. I was able to get them passing by using
`assertEventually` to repeatedly check until the condition passes, failing once
a max timeout is reached.

(TODO: Remove this inlined version of `assertEventually` and pull it in from
mongochangestream-testing instead.)

Results:

 ? src/convertSchema.test.ts (9)
 ? src/syncData.test.ts (2) 22074ms

 Test Files  2 passed (2)
      Tests  11 passed (11)
   Start at  13:55:28
   Duration  22.58s (transform 70ms, setup 0ms, collect 431ms, tests 22.09s, environment 0ms, prepare 80ms)